pythonapi/requirements.txt
fastapi
redis
types-redis
uvicorn
requests
pandas
load_dotenv